[Chronik] Refactor bridge: Split into `chronik-bridge` and `chronik-lib`
Summary:
Currently, all the ffi code is in `chronik-ffi`. However, other crates might want to depend on the ffi code.

However, `chronik-ffi` is the staticlib that is linked to bitcoind at the end, it'll be the "root" crate depending on all the crates comprising Chronik.

For that reason, other crates cannot depend on `chronik-ffi`, as it will create a circular reference, which is not allowed for crates.

Say a crate `chronik-log` wants to reference `log_println`, defined in `chronik-ffi`. Also, `chronik-ffi` wants to use the logging defined in `chronik-log`.

It would create circular dependencies: `chronik-log` -> `chronik-ffi` -> `chronik-log`.

Therefore, we have to split the ffi into two parts:
1. `chronik-bridge`: This defines shared types and functions defined in C++, e.g. `log_println`.
2. `chronik-lib`: This defines all the functions that C++ will need to call, like `setup_bridge` and methods from `CValidationInterface`.

This way, there's no circular dependencies: `chronik-bridge` -> `chronik-log` -> `chronik-lib`.

Test Plan:
1. `ninja`
2. `./src/bitcoind -regtest` prints "Starting Chronik..." as before

Reviewers: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11909