[Chronik] Shut down node after errors in merge operators gracefully
Summary:
Currently, when the merge operator `full_merge_concat_trim` failes, it panics. However, this may cause issues with corrupting the db or even the node.

Therefore, instead of panicing, we log the error message, and then return None from the merge function, which will cause RocksDB to return a `Corruption` error, which then shuts down the node gracefully.

Example:
```
MERGE ERROR: Bad usage of merge: Unknown prefix 45, expected C or T: 4500
RocksDB error: Corruption: Merge operator failed
```

Directly using log! in chronik-db, however, causes linker issues when running `cargo test`: When `chronik-indexer` tests are compiled, the dependency `chronik-db` is *not* compiled with `cfg(test)` on, because it's a dependency. Therefore, it will compile with the LogPrint function from the node, which in undefined during `cargo test`.

To avoid this, we set logging functions externally (during `setup_chronik`) to a `OnceLock` (static field that can only be set once), which are left to println! during `cargo test`.

Test Plan:
1. `ninja && ninja check-crates && ./test/functional/test_runner.py chronik_script_history`
2. Change `CONCAT` in `insert` to `b'E'` and run the functional test again, now it will gracefully shut down the node with an error.

Reviewers: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D14408