[Chronik] Enable SLP/ALP index, add protobuf messages & fields
Summary:
- Enable the SLP/ALP index unconditionally
- Add protobuf messages & fields for SLP/ALP
- Add `TxTokenData` to read the required mempool/DB data and output the protobuf

Note that there's no ability to disable the token index, or reindex it stand-alone. This is because the specifics can be unwieldy; e.g. RocksDB requires us to
specify the column families to open the DB with, we'll need to conditionally index as well as read from the DB. Also, we might want to add more indices or even a plugin
interface in the future, in this case it would be better to have a generalized system to enable/disable/selectively reindex certain indices.

Test Plan: `ninja check-functional`

Reviewers: Fabien, bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D15105