Update wallet_multiwallet.py for descriptor
Summary:
This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19077 | core#19077]]

Note that none of these changes was actually necessary to make this test pass with `--descriptors`, but my main goal with this backport is to make subsequent backports easier, so pick as many changes as possible.
The change to BitcoinTestFramework.default_wallet_name and the "Unable to obtain an exclusive lock" error message are not applicable due to missing sqlite related backports.

Test Plan:
`test/functional/test_runner.py wallet_multiwallet`
Check that `wallet_multiwallet.py --descriptors` passes

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15142