[Chronik] Fix: Missing entry for "burned" NFT1 GROUP token for NFT1 CHILD GENESIS txs
Summary:
NFT1 CHILD GENESIS txs require a NFT1 GROUP token to be "burned" in the first input.

Currently, we don't create an entry for those, which results in a tx with token inputs that don't have a corresponding entry, which we don't want.

This diff uses the bare_burn map to add an entry to fix this.

Test Plan: `cargo test -p bitcoinsuite-slp --test test_verify_slp_genesis`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D15177