[Chronik] Use `merge_cf` in `reverse_lookup.rs` to speed up indexing.
Summary:
`TxWriter` is surprizingly slow considering what it does is very simple, however, the bottleneck is `LookupIndex`, which in `insert_pairs` reads from the DB before updating, which is very slow.

Instead, use merge_cf like in D14430, which avoids that roundrip, and RocksDB is able to combine a read+write efficiently for us.

Depends on D15547.

Test Plan: `ninja check-crates && ninja check-functional`

Reviewers: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D15548