[Chronik] Add `TxNumCache` to speed up `prepare_indexed_txs`
Summary:
Currently, `prepare_indexed_txs` is still one of the slowest parts of the indexer due to the disk access, even when using batched queries.

We add a `TxNumCache`, which simply stores the last `depth_blocks` worth of tx nums in a VecDeque of HashMaps. The latest block of txids is dropped off at the end, and the new block is added at the beginning.

Benchmarks TBD.

Test Plan: `ninja check-functional`

Reviewers: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D15619