[Chronik] Add Docusaurus documentation
Summary:
This provides a full reference on how to use Chronik, from downloading, to building, to setup and how to use `chronik-client`.

This code is based on the auto-generated template tutorial for docusaurus.

It uses mdx, which is a mix of Markdown and JSX, which gives us the ability to include Live Code so the user can play around with the code examples right there.

Test Plan:
1. in `modules/chronik-client`: `npm install`
2. in `doc/chronik`: `npm install && npm serve`
3. Should open browser automatically, go through all pages and check if the information is correct

Reviewers: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D15665