[ecash-agora] Add `AgoraPartial.script` to build the Script for partial offers
Summary:
Add a function building the SLP version of the script for offers that can be partially accepted.

This doesn't add a way of spending them yet, and tests are more to "pin" the script code (which is important, so future refactors don't change the script code), than to actually test if they work as intended. All of which will be added in a future diff.

Depends on D16735.

Test Plan: npm run test

Reviewers: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D16743