[Chronik] Automatically fix MINT VAULT txs
This is part of the DB upgrade to version 13.

It seems that at this moment of time, this only upgrades a single tx, 47f663238196a8376dffeb867727b7794f1a86e97d7401dd141ac8c072f91c54. However, if someone is updating the DB at a later stage, this code will handle this gracefully (and there seems indication that more txs of this form will come in the future).